moved booking service hook to helpers
