+++ src/booking/stages/adding-extensions.tsx
@@ -3,7 +3,7 @@ import { useActor } from "@xstate/react";
 import Button from "src/components/button";
 import Modal from "src/components/modal";
 import styles from "src/booking/stages/adding-extensions.module.scss";
-import { useBookingService } from "src/booking/useBookingService";
+import { useBookingService } from "src/helpers";
 
 const AddingExtensions = ({ service, setShowModal }) => {
   const { bookingService } = useBookingService();
