--- src/booking/useBookingService.js
@@ -1,4 +1,0 @@
-import React from "react";
-import { BookingContext } from "src/booking/booking-state";
-
-export const useBookingService = () => React.useContext(BookingContext);
