+++ src/helpers.tsx
@@ -1,5 +1,6 @@
+import { useState, useEffect, useContext } from "react";
 import { AppImage } from "src/types/image";
-import { useState, useEffect } from "react";
+import { BookingContext } from "src/booking/booking-state";
 
 const MOBILE = 768;
 
@@ -28,3 +29,5 @@ export const useBreakpoints = () => {
 
   return { mobile };
 };
+
+export const useBookingService = () => useContext(BookingContext);
